Revert "refactor: remove mult index and external type file"
