+++ src/components/Header/Header.types.ts
@@ -1,0 +1,3 @@
+export type HeaderProps = {
+  clearTasks: () => void;
+};
