+++ src/components/Tasks/Tasks.types.ts
@@ -1,0 +1,6 @@
+import { Dispatch } from "react";
+
+export type TasksProps = {
+  tasks: string[];
+  setTasks: Dispatch<string[]>;
+};
