+++ src/components/Tasks/index.ts
@@ -1,0 +1,1 @@
+export { Tasks } from "./Tasks";
