+++ src/components/index.ts
@@ -1,4 +1,4 @@
-export { Footer } from "./Footer/Footer";
-export { Header } from "./Header/Header";
-export { Logo } from "./Logo/Logo";
-export { Tasks } from "./Tasks/Tasks";
+export { Footer } from "./Footer";
+export { Header } from "./Header";
+export { Logo } from "./Logo";
+export { Tasks } from "./Tasks";
